chore: move api client to public pkg
