+++ tools/codegen/client.cfg.yaml
@@ -1,4 +1,4 @@
 package: api
 generate:
   client: true
-output: ../../internal/api/client.gen.go
+output: ../../pkg/api/client.gen.go
